+++ runtime/src/capi/Player.h
@@ -29,12 +29,12 @@ extern "C"
     EXPORT void Player_Despawn(void *p);
     EXPORT void Player_SetModel(void *p, unsigned long model);
     EXPORT unsigned long Player_GetModel(void *p);
-    EXPORT unsigned long Player_GetHealth(void *p);
-    EXPORT void Player_SetHealth(void *p, unsigned long health);
+    EXPORT unsigned int Player_GetHealth(void *p);
+    EXPORT void Player_SetHealth(void *p, unsigned int health);
     EXPORT bool Player_HasWeaponComponent(void *p, unsigned long weapon, unsigned long component);
     //EXPORT alt::Array<unsigned int> Player_GetCurrentWeaponComponents(void *p);
-    EXPORT unsigned long Player_GetWeaponTintIndex(void *p, unsigned long weapon);
-    EXPORT unsigned long Player_GetCurrentWeaponTintIndex(void *p);
+    EXPORT unsigned int Player_GetWeaponTintIndex(void *p, unsigned long weapon);
+    EXPORT unsigned int Player_GetCurrentWeaponTintIndex(void *p);
     EXPORT unsigned long Player_GetCurrentWeapon(void *p);
     EXPORT bool Player_IsDead(void *p);
     EXPORT bool Player_IsJumping(void *p);
@@ -42,14 +42,14 @@ extern "C"
     EXPORT bool Player_IsAiming(void *p);
     EXPORT bool Player_IsShooting(void *p);
     EXPORT bool Player_IsReloading(void *p);
-    EXPORT unsigned long Player_GetArmour(void *p);
-    EXPORT void Player_SetArmour(void *p, unsigned long armour);
+    EXPORT unsigned int Player_GetArmour(void *p);
+    EXPORT void Player_SetArmour(void *p, unsigned int armour);
     EXPORT float Player_GetMoveSpeed(void *p);
     EXPORT Position Player_GetAimPos(void *p);
     EXPORT Rotation Player_GetHeadRotation(void *p);
     EXPORT bool Player_IsInVehicle(void *p);
     EXPORT void * Player_GetVehicle(void *p);
-    EXPORT unsigned long Player_GetSeat(void *p);
+    EXPORT unsigned int Player_GetSeat(void *p);
     EXPORT void * Player_GetEntityAimingAt(void *p);
     EXPORT Position Player_GetEntityAimOffset(void *p);
     EXPORT bool Player_IsFlashlightActive(void *p);
@@ -60,14 +60,14 @@ extern "C"
     EXPORT unsigned long long Player_GetHwidHash(void *p);
     EXPORT unsigned long long Player_GetHwidExHash(void *p);
     EXPORT const char* Player_GetAuthToken(void *p);
-    EXPORT void Player_SetMaxArmour(void *p, unsigned long armour);
+    EXPORT void Player_SetMaxArmour(void *p, unsigned int armour);
     EXPORT void Player_SetCurrentWeapon(void *p, unsigned long weapon);
-    EXPORT void Player_SetWeaponTintIndex(void *p, unsigned long weapon, unsigned long tintIndex);
+    EXPORT void Player_SetWeaponTintIndex(void *p, unsigned long weapon, unsigned int tintIndex);
     EXPORT void Player_AddWeaponComponent(void *p, unsigned long weapon, unsigned long component);
     EXPORT void Player_RemoveWeaponComponent(void *p, unsigned long weapon, unsigned long component);
     EXPORT void Player_ClearBloodDamage(void *p);
-    EXPORT void Player_SetMaxHealth(void *p, unsigned long health);
-    EXPORT void Player_GiveWeapon(void *p, unsigned long weapon, unsigned long ammo, bool selectWeapon);
+    EXPORT void Player_SetMaxHealth(void *p, unsigned int health);
+    EXPORT void Player_GiveWeapon(void *p, unsigned long weapon, long ammo, bool selectWeapon);
     EXPORT void Player_RemoveWeapon(void *p, unsigned long weapon);
     EXPORT void Player_RemoveAllWeapons(void *p);
     EXPORT void Player_SetDateTime(void *p, int day, int month, int year, int hour, int minute, int second);
@@ -76,16 +76,16 @@ extern "C"
     /**
      * Cloth and Prop Getters
      * */
-    EXPORT void Player_SetClothes(void *p, unsigned long component, unsigned long drawable, unsigned long texture, unsigned long palette);
-    EXPORT void Player_SetDlcClothes(void *p, unsigned long component, unsigned long drawable, unsigned long texture, unsigned long palette, unsigned long dlc);
-    EXPORT void Player_SetProps(void *p, unsigned long component, unsigned long drawable, unsigned long texture);
-    EXPORT void Player_SetDlcProps(void *p, unsigned long component, unsigned long drawable, unsigned long texture, unsigned long dlc);
-    EXPORT void Player_ClearProps(void *p, unsigned long component);
+    EXPORT void Player_SetClothes(void *p, unsigned int component, unsigned int drawable, unsigned int texture, unsigned int palette);
+    EXPORT void Player_SetDlcClothes(void *p, unsigned int component, unsigned int drawable, unsigned int texture, unsigned int palette, unsigned long dlc);
+    EXPORT void Player_SetProps(void *p, unsigned int component, unsigned int drawable, unsigned int texture);
+    EXPORT void Player_SetDlcProps(void *p, unsigned int component, unsigned int drawable, unsigned int texture, unsigned long dlc);
+    EXPORT void Player_ClearProps(void *p, unsigned int component);
     EXPORT bool Player_IsEntityInStreamingRange(void *p, void *entity);
-    EXPORT unsigned long Player_GetMaxHealth(void *p);
-    EXPORT unsigned long Player_GetMaxArmour(void *p);
+    EXPORT unsigned int Player_GetMaxHealth(void *p);
+    EXPORT unsigned int Player_GetMaxArmour(void *p);
     EXPORT void Player_Detach(void *p);
-    EXPORT void Player_AttachToEntity(void *p, void *e, unsigned long otherBoneIndex, unsigned long myBoneIndex, Position position, Rotation rotation, bool collision, bool noFixedRotation);
+    EXPORT void Player_AttachToEntity(void *p, void *e, int otherBoneIndex, int myBoneIndex, Position position, Rotation rotation, bool collision, bool noFixedRotation);
     EXPORT void Player_SetVisible(void *p, bool toggle);
     EXPORT bool Player_GetVisible(void *p);
     EXPORT unsigned long Player_GetID(void *p);
